eachRecord: always enumerate records in adt file order
Fixes intermittent/non-deterministic unit test failures.
The unit tests are written assuming records are returned in a fixed
order. But the eachRecord function is non-deterministic in the order
in which records will be sent to the iterator callback. This change
ensures eachRecord always returns records in the order they appear
in the adt file.